requirements.txt
appdirs==1.4.4
apturl==0.5.2
argcomplete==3.1.1
asttokens==2.1.0
autopep8==2.0.0
backcall==0.2.0
bcrypt==3.2.0
black==23.3.0
blinker==1.4
Brlapi==0.8.3
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dacite==1.6.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fasteners==0.14.1
filelock==3.9.0
future==0.18.2
googlemaps==4.10.0
greenlet==1.1.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.17.0
ipython==8.6.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jupyter_client==7.4.4
jupyter_core==4.11.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.0
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.31
nest-asyncio==1.5.6
netifaces==0.11.0
numpy==1.22.3
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.9
packaging==23.1
pandas==1.4.1
paramiko==2.9.3
parso==0.8.3
pathspec==0.11.1
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2023.2.18
platformdirs==3.0.0
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycodestyle==2.9.1
pycups==2.0.1
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pypng==0.20220715.0
pyRFC3339==1.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.21.0
python-escpos==2.2.0
pytz==2021.3
pyusb==1.2.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
qrcode==7.4.2
reportlab==3.6.8
requests==2.27.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==1.4.32
stack-data==0.6.0
systemd-python==234
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.8
usb-creator==0.3.7
virtualenv==20.19.0
virtualenv-clone==0.5.7
wadllib==1.3.6
wcwidth==0.2.5
wkhtmltopdf==0.2
xdg==5
xero-python==1.20.0
xkit==0.0.0
zipp==1.0.0
